+++ src/ApplicationSettings.h
@@ -93,6 +93,7 @@ public:
     DEFINE_SETTING(ShowTabBar, showTabBar, bool)
     DEFINE_SETTING(ShowStatusBar, showStatusBar, bool)
     DEFINE_SETTING(CenterSearchDialog, centerSearchDialog, bool)
+    DEFINE_SETTING(Theme, theme, int)
 
     DEFINE_SETTING(TabsClosable, tabsClosable, bool)
     DEFINE_SETTING(ExitOnLastTabClosed, exitOnLastTabClosed, bool)